Move spike train features into a separate module
